Mock services using jest.mock
